+++ src/KristofferStrube.Blazor.Streams/BaseJSStreamableWrapper.cs
@@ -1,0 +1,33 @@
+﻿using Microsoft.JSInterop;
+
+namespace KristofferStrube.Blazor.Streams;
+
+public abstract class BaseJSStreamableWrapper : Stream, IAsyncDisposable
+{
+    public readonly IJSObjectReference JSReference;
+    protected readonly Lazy<Task<IJSObjectReference>> helperTask;
+    protected readonly IJSRuntime jSRuntime;
+
+    /// <summary>
+    /// Constructs a wrapper instance for an equivalent JS instance.
+    /// </summary>
+    /// <param name="jSRuntime">An <see cref="IJSRuntime"/> instance.</param>
+    /// <param name="jSReference">A JS reference to an existing JS instance that should be wrapped..</param>
+    internal BaseJSStreamableWrapper(IJSRuntime jSRuntime, IJSObjectReference jSReference)
+    {
+        helperTask = new(() => jSRuntime.GetHelperAsync());
+        JSReference = jSReference;
+        this.jSRuntime = jSRuntime;
+    }
+
+    public new async ValueTask DisposeAsync()
+    {
+        await base.DisposeAsync();
+        if (helperTask.IsValueCreated)
+        {
+            IJSObjectReference module = await helperTask.Value;
+            await module.DisposeAsync();
+        }
+        GC.SuppressFinalize(this);
+    }
+}