+++ src/KristofferStrube.Blazor.Streams/ReadableStream/ReadableStream.cs
@@ -5,12 +5,8 @@ namespace KristofferStrube.Blazor.Streams;
 /// <summary>
 /// <see href="https://streams.spec.whatwg.org/#rs-class-definition">Streams browser specs</see>
 /// </summary>
-public partial class ReadableStream : IAsyncDisposable
+public partial class ReadableStream : BaseJSStreamableWrapper
 {
-    public readonly IJSObjectReference JSReference;
-    protected readonly Lazy<Task<IJSObjectReference>> helperTask;
-    protected readonly IJSRuntime jSRuntime;
-
     /// <summary>
     /// Constructs a wrapper instance for a given JS Instance of a <see cref="ReadableStream"/>.
     /// </summary>
@@ -70,12 +66,7 @@ public partial class ReadableStream : IAsyncDisposable
     /// </summary>
     /// <param name="jSRuntime">An <see cref="IJSRuntime"/> instance.</param>
     /// <param name="jSReference">A JS reference to an existing <see cref="ReadableStream"/>.</param>
-    internal ReadableStream(IJSRuntime jSRuntime, IJSObjectReference jSReference)
-    {
-        helperTask = new(() => jSRuntime.GetHelperAsync());
-        JSReference = jSReference;
-        this.jSRuntime = jSRuntime;
-    }
+    internal ReadableStream(IJSRuntime jSRuntime, IJSObjectReference jSReference) : base(jSRuntime, jSReference) { }
 
     /// <summary>
     /// Indicates whether the stream already has a reader.
@@ -167,16 +158,5 @@ public partial class ReadableStream : IAsyncDisposable
             new ReadableStream(jSRuntime, await helper.InvokeAsync<IJSObjectReference>("elementAt", jSArray, 0)),
             new ReadableStream(jSRuntime, await helper.InvokeAsync<IJSObjectReference>("elementAt", jSArray, 1))
             );
-    }
-
-    public new async ValueTask DisposeAsync()
-    {
-        await base.DisposeAsync();
-        if (helperTask.IsValueCreated)
-        {
-            IJSObjectReference module = await helperTask.Value;
-            await module.DisposeAsync();
-        }
-        GC.SuppressFinalize(this);
     }
 }